build: moving local function to streamlit tools or utilities
