+++ streamlit_toolkit/tools.py
@@ -1,6 +1,8 @@
 import os
 import time
 import streamlit as st
+import tkinter as tk
+from tkinter import filedialog
 
 import modules.utilities as util
 import modules.INFO as INFO
@@ -41,6 +43,166 @@ def download_as():
             file_name=f'log_{INFO.SESSION_TIME}.txt',
             mime='text/plain'
         )
+
+
+def save(content, path, page='', json_value: dict = None):
+    if json_value is None:
+        json_value = []
+    save_but = st.button(_('💾Save'))
+    if save_but:
+        util.write_file(content, path)
+        st.success(_('✅File saved!'))
+        # write to json file
+        if page == '💽Brain Memory':
+            util.update_json(INFO.BRAIN_MEMO, 'delimiter', json_value['delimiter'])
+            util.update_json(INFO.BRAIN_MEMO, 'append_mode', json_value['append_mode'])
+            util.update_json(INFO.BRAIN_MEMO, 'force_mode', json_value['force_mode'])
+            util.update_json(INFO.BRAIN_MEMO, 'advanced_mode', json_value['advanced_mode'])
+            util.update_json(INFO.BRAIN_MEMO, 'filter_info', json_value['filter_info'])
+            util.update_json(INFO.BRAIN_MEMO, 'filter_row_count', json_value['filter_row_count'])
+        time.sleep(1)
+        # refresh page
+        st.experimental_rerun()
+
+
+def match_logic(operator, filter_val, value):
+    if operator == 'IS':
+        return filter_val == value
+    elif operator == 'IS NOT':
+        return filter_val != value
+    elif operator == 'CONTAINS':
+        return filter_val in value
+    elif operator == 'NOT CONTAINS':
+        return filter_val not in value
+    elif operator == 'MORE THAN':
+        # check if value is float
+        if not value.isnumeric():
+            return False
+        return float(filter_val) < float(value)
+    elif operator == 'LESS THAN':
+        # check if value is float
+        if not value.isnumeric():
+            return False
+        return float(filter_val) > float(value)
+    elif operator == 'MORE THAN OR EQUAL':
+        # check if value is float
+        if not value.isnumeric():
+            return False
+        return float(filter_val) <= float(value)
+    elif operator == 'LESS THAN OR EQUAL':
+        # check if value is float
+        if not value.isnumeric():
+            return False
+        return float(filter_val) >= float(value)
+    else:
+        return False
+
+
+def select_directory():
+    root = tk.Tk()
+    root.withdraw()
+    # make sure the dialog is on top of the main window
+    root.attributes('-topmost', True)
+    directory = filedialog.askdirectory(initialdir=os.getcwd(), title=_('Select Note Directory'))
+    return directory
+
+
+def match_fields(pages: list, filter_datas: list[dict]):
+    filtered_contents = []
+    for page in pages:
+        fields = util.extract_frontmatter(page, delimiter='---')
+
+        found_data = []
+
+        for field in fields:
+            if field == '':
+                continue
+            found_key, found_value = field.split(':')
+            found_key = found_key.strip()
+            found_value = found_value.strip()
+
+            found_data.append({
+                'key': found_key,
+                'value': found_value
+            })
+
+        found_match = []
+        for data in filter_datas:
+            for found in found_data:
+                data_key = data['key'].lower()
+                data_val = data['value'].lower()
+                found_key = found['key'].lower()
+                found_val = found['value'].lower()
+                if data_key == found_key:
+                    if match_logic(data['logic'], data_val, found_val):
+                        # found single match
+                        found_match.append(True)
+
+        # if all match
+        if found_match.count(True) == len(filter_datas):
+            filtered_contents.append(page)
+
+    combined_contents = '\n\n\n\n'.join(filtered_contents)
+    return combined_contents
+
+
+def add_filter(num, val_filter_key, val_filter_logic, val_filter_val):
+    # filters
+    col1, col2, col3 = st.columns(3)
+    with col1:
+        filter_key = st.text_input(f'Key{num}', placeholder='Key', value=val_filter_key)
+    with col2:
+        options = ['CONTAINS',
+                   'NOT CONTAINS',
+                   'IS',
+                   'IS NOT',
+                   'MORE THAN',
+                   'LESS THAN',
+                   'MORE THAN OR EQUAL',
+                   'LESS THAN OR EQUAL']
+        default_index = util.get_index(options, val_filter_logic, 0)
+        logic_select = st.selectbox(f'Logic{num}', options, index=default_index)
+    with col3:
+        if isinstance(val_filter_val, int):
+            val_filter_val = "{:02}".format(val_filter_val)
+        filter_val = st.text_input(f'value{num}', placeholder='Value', value=val_filter_val)
+    return filter_key, logic_select, filter_val
+
+
+def filter_data(pages: list, add_filter_button, del_filter_button):
+    init_filter_infos = util.read_json_at(INFO.BRAIN_MEMO, 'filter_info')
+
+    filter_datas = []
+    if add_filter_button:
+        st.session_state['FILTER_ROW_COUNT'] += 1
+    if del_filter_button:
+        st.session_state['FILTER_ROW_COUNT'] -= 1
+    if st.session_state['FILTER_ROW_COUNT'] >= 1:
+        for i in range(st.session_state['FILTER_ROW_COUNT'] + 1):
+            try:
+                init_info = init_filter_infos[i - 1]
+                init_key = init_info['key']
+                init_logic = init_info['logic']
+                init_val = init_info['value']
+            except IndexError:
+                init_key = ''
+                init_logic = 'CONTAINS'
+                init_val = ''
+            except KeyError:
+                init_key = ''
+                init_logic = 'CONTAINS'
+                init_val = ''
+
+            if i == 0:
+                continue
+            # add filter
+            filter_key, logic_select, filter_val = add_filter(i, init_key, init_logic, init_val)
+            data = {'key': filter_key, 'logic': logic_select, 'value': filter_val}
+            filter_datas.append(data)
+
+    # filter data
+    filtered_contents = match_fields(pages, filter_datas)
+    return filtered_contents, filter_datas
 
 
 def process_response(query, target_model, prompt_file: str, data: GPT.model.param):
